WIP refactor download activity
