+++ apps/vessel-history/features/event-filters/EventFiltersButton.tsx
@@ -2,7 +2,6 @@ import { Fragment } from 'react'
 import cx from 'classnames'
 import { useSelector } from 'react-redux'
 import { useTranslation } from 'react-i18next'
-import DownloadActivityCsv from 'feature/download-activity-csv/download-activity-csv'
 import { Button, Icon, IconButton, ButtonType } from '@globalfishingwatch/ui-components'
 import { selectFilterUpdated } from 'features/event-filters/filters.selectors'
 import FiltersLabel from 'features/filters-label/filters-label'
@@ -15,9 +14,6 @@ interface ButtonProps {
   className?: string
   type?: ButtonType
   onClick?: () => void
-  onDownloadAllActivityCsv?: () => void
-  onDownloadFilteredActivityCsv?: () => void
-  onReadmeClick?: () => void
 }
 
 const EventFiltersButton: React.FC<ButtonProps> = ({ className, ...props }): React.ReactElement => {
@@ -44,17 +40,6 @@ const EventFiltersButton: React.FC<ButtonProps> = ({ className, ...props }): Rea
           <FiltersLabel filters={filters} />
         </Button>
       )}
-      {!currentProfileIsInsurer &&
-        (props.onDownloadAllActivityCsv !== undefined ||
-          props.onDownloadFilteredActivityCsv !== undefined ||
-          props.onReadmeClick !== undefined) && (
-          <DownloadActivityCsv
-            filtersApplied
-            onDownloadAllActivityCsv={props.onDownloadAllActivityCsv}
-            onDownloadFilteredActivityCsv={props.onDownloadFilteredActivityCsv}
-            onReadmeClick={props.onReadmeClick}
-          />
-        )}
     </Fragment>
   )
 }
