+++ apps/vessel-history/features/profile/filters/ActivityFilters.tsx
@@ -1,24 +1,21 @@
 import { Fragment, useCallback, useState } from 'react'
 import { event as uaEvent } from 'react-ga'
 import { useTranslation } from 'react-i18next'
+import DownloadActivity from 'feature/download-activity/download-activity'
+import { useSelector } from 'react-redux'
 import EventFilters from 'features/event-filters/EventFilters'
 import DataAndTerminology from 'features/data-and-terminology/DataAndTerminology'
 import EventFiltersButton from 'features/event-filters/EventFiltersButton'
+import { selectFilterUpdated } from 'features/event-filters/filters.selectors'
 import ActivityDataAndTerminology from '../components/activity/ActivityDataAndTerminology'
+import { selectCurrentUserProfileHasInsurerPermission } from '../profile.selectors'
 import styles from './ActivityFilters.module.css'
 
-interface ActivityFiltersProps {
-  onDownloadCsv?: () => void
-  onDownloadAllActivityCsv?: () => void
-  onDownloadFilteredActivityCsv?: () => void
-  onReadmeClick?: () => void
-}
+interface ActivityFiltersProps {}
 
-const ActivityFilters: React.FC<ActivityFiltersProps> = ({
-  onDownloadAllActivityCsv,
-  onDownloadFilteredActivityCsv,
-  onReadmeClick,
-}: ActivityFiltersProps): React.ReactElement => {
+const ActivityFilters: React.FC<ActivityFiltersProps> = (
+  props: ActivityFiltersProps
+): React.ReactElement => {
   const { t } = useTranslation()
   const [isModalOpen, setIsOpen] = useState(false)
   const setModalOpen = useCallback((isOpen) => {
@@ -28,6 +25,18 @@ const ActivityFilters: React.FC<ActivityFiltersProps> = ({
       label: JSON.stringify({ tab: 'ACTIVITY' }),
     })
     setIsOpen(isOpen)
+  }, [])
+  const filtersApplied = useSelector(selectFilterUpdated)
+  const currentProfileIsInsurer = useSelector(selectCurrentUserProfileHasInsurerPermission)
+
+  const onDownloadAllActivityCsv = useCallback(() => {
+    // TODO: Add google analytics tracking for this event
+  }, [])
+  const onDownloadFilteredActivityCsv = useCallback(() => {
+    // TODO: Add google analytics tracking for this event
+  }, [])
+  const onReadmeClick = useCallback(() => {
+    // TODO: Add google analytics tracking for this event
   }, [])
 
   return (
@@ -49,10 +58,15 @@ const ActivityFilters: React.FC<ActivityFiltersProps> = ({
         <EventFiltersButton
           type="secondary"
           onClick={() => setModalOpen(true)}
-          onDownloadAllActivityCsv={onDownloadAllActivityCsv}
-          onDownloadFilteredActivityCsv={onDownloadFilteredActivityCsv}
-          onReadmeClick={onReadmeClick}
         ></EventFiltersButton>
+        {!currentProfileIsInsurer && (
+          <DownloadActivity
+            filtersApplied={filtersApplied}
+            onDownloadAllActivityCsv={onDownloadAllActivityCsv}
+            onDownloadFilteredActivityCsv={onDownloadFilteredActivityCsv}
+            onReadmeClick={onReadmeClick}
+          />
+        )}
       </div>
     </Fragment>
   )
